Fix sipix camlib compilation warnings.
* blink2.c:
  * file_list_func(...): Change variable 'i' type to unsigned to be used in for loop with unsigned compare. Also add cast to 'ret' variable to unsigned as this does not change behavior and 'ret' is guaranteed to be >= GP_OK.
  * get_file_func(...): Same as above. Change 'image_no >= numpics' comparison. Functionally has the same behavior. Change variable 'j' type to unsigned for unsigned 'for' loop comparison. Finally change 'curread > len' comparison to suppress signed-unsigned comparison warning.
* web2.c:
  * web2_getpicture(...): Add unsigned cast to 'toread' variable. This does not change behavior and variable 'toread' is guaranteed to be positive integer.

+++ camlibs/sipix/web2.c
@@ -181,7 +181,7 @@ web2_getpicture(GPPort *port, GPContext *context, CameraFile *file)
 
     while (curread < size) {
 	int toread = size-curread;
-	if (toread > sizeof(xbuf))
+	if ((unsigned int)toread > sizeof(xbuf))
 	    toread = sizeof(xbuf);
 	ret = gp_port_read(port, xbuf, toread);
 	if (ret < GP_OK)
